Moved the mixin to RegionAceAccess
