+++ worlds/oot_soh/RegionAgeAccess.py
@@ -1,105 +1,80 @@
 from typing import Dict, NamedTuple, cast, TYPE_CHECKING
-from BaseClasses import CollectionState, Entrance, Region
-from .Enums import Regions
+from collections import deque
+from BaseClasses import CollectionState, MultiWorld
+from worlds.AutoWorld import LogicMixin
+from .Enums import Regions, Ages
+import copy
 
-if TYPE_CHECKING:
-    from . import SohWorld
+class SohAgeLogic(LogicMixin):
+    def init_mixin(self, parent: MultiWorld):
+        game = "Ship of Harkinian"
+        self._soh_stale = {player: True for player in parent.worlds.keys()
+                            if parent.worlds[player].game == game}
+        players = parent.get_game_groups(game) + parent.get_game_players(game)
+        self._soh_child_reachable_regions = {player: set() for player in players}
+        self._soh_adult_reachable_regions = {player: set() for player in players}
+        self._soh_child_blocked_regions = {player: set() for player in players}
+        self._soh_adult_blocked_regions = {player: set() for player in players}
+        self._soh_age = {player: Ages.null for player in players}
 
-class SohEntranceData(NamedTuple):
-    can_pass_as_child: bool = True
-    can_pass_as_adult: bool = True
+    def copy_mixin(self, ret) -> CollectionState:
+        ret._soh_stale = {player: stale for player, stale in self._soh_stale.items()}
+        ret._soh_child_reachable_regions = {player: copy.copy(regions) for player, regions in self._soh_child_reachable_regions.items()}
+        ret._soh_adult_reachable_regions = {player: copy.copy(regions) for player, regions in self._soh_adult_reachable_regions.items()}
+        ret._soh_child_blocked_regions = {player: copy.copy(regions) for player, regions in self._soh_child_blocked_regions.items()}
+        ret._soh_adult_blocked_regions = {player: copy.copy(regions) for player, regions in self._soh_adult_blocked_regions.items()}
+        ret._soh_age = {player: age for player, age in self._soh_age.items()}
+        return ret
+    
+    def _soh_invalidate(self, player):
+        self._soh_child_reachable_regions[player] = set()
+        self._soh_adult_reachable_regions[player] = set()
+        self._soh_child_blocked_regions[player] = set()
+        self._soh_adult_blocked_regions[player] = set()
+        self._soh_stale[player] = True
 
-entrance_data_table: Dict[str, SohEntranceData] = {}
-child_access_table: Dict[str, bool] = {}
-adult_access_table: Dict[str, bool] = {}
+    def _soh_update_age_reachable_regions(self, player):
+        self._soh_stale[player] = False
+        for age in [Ages.CHILD, Ages.ADULT]:
+            self._soh_age[player] = age
+            start = self.multiworld.get_region(Regions.ROOT, player)
+            
+            if age == Ages.CHILD:
+                reachable = self._soh_child_reachable_regions[player]
+                blocked = self._soh_child_blocked_regions[player]
+                queue = deque(self._soh_child_blocked_regions[player])
+            else:
+                reachable = self._soh_adult_reachable_regions[player]
+                blocked = self._soh_adult_blocked_regions[player]
+                queue = deque(self._soh_adult_blocked_regions[player])
 
-def reset_age_access(start_as_adult: bool = False):
-    for region in Regions:
-        child_access_table[region.value] = False
-        adult_access_table[region.value] = False
+            # init on first call
+            if start not in reachable:
+                reachable.add(start)
+                blocked.update(start.exits)
+                queue.extend(start.exits)
 
-    child_access_table[Regions.ROOT.value] = not start_as_adult
-    adult_access_table[Regions.ROOT.value] = start_as_adult
-
-reset_age_access()
-
-# Returns whether a change was made
-def update_age_access(world: "SohWorld", state: CollectionState):
-    # Spread from Menu
-    # If ToT accessible, spread from ToT
-    # Any access at our starting age that ToT access opens up would be covered in the ToT floodfill, since if it mattered, we wouldn't reach that point
-    just_found_tot = _spread_age_access(world, state, Regions.ROOT.value)
-    if just_found_tot: # As said, if we just found ToT, we probably need to do a floodfill from there
-        _spread_age_access(world, state, Regions.TEMPLE_OF_TIME.value)
-
-# Returns whether ToT was just found
-def _spread_age_access(world: "SohWorld", state: CollectionState, root: str) -> bool:
-    region_list = [root]
-    just_found_tot = False
-    while len(region_list) > 0:
-        region = world.get_region(region_list.pop())
-        if child_access_table[region.name]:
-            for exit in region.exits:
-                # The last part breaks an infinite loop; this way, any cycle in the graph is broken as the boolean will have been made true
-                if exit.connected_region is None or (exit.name in entrance_data_table and not entrance_data_table[exit.name].can_pass_as_child) \
-                    or (exit.connected_region.name in child_access_table and child_access_table[exit.connected_region.name]):
+            # run breadth first search
+            while queue:
+                connection = queue.popleft()
+                new_region = connection.connected_region
+                if new_region is None:
                     continue
-                if state.can_reach_entrance(exit.name, world.player):
-                    next_name = exit.connected_region.name
-                    child_access_table[next_name] = True
-                    if next_name not in region_list:
-                        region_list.append(next_name)
-                    if next_name == Regions.TEMPLE_OF_TIME.value: # If we just got access to ToT, we need to reset the floodfill to cover the adult side
-                        # If we are in this clause, we first gained access as child, which must have been our starting age
-                        adult_access_table[next_name] = True
-                        just_found_tot = True
-        if adult_access_table[region.name]:
-            for exit in region.exits:
-                # The last part breaks an infinite loop; this way, any cycle in the graph is broken as the boolean will have been made true
-                if exit.connected_region is None or (exit.name in entrance_data_table and not entrance_data_table[exit.name].can_pass_as_adult) \
-                    or (exit.connected_region.name in adult_access_table and adult_access_table[exit.connected_region.name]):
-                    continue
-                if state.can_reach_entrance(exit.name, world.player):
-                    next_name = exit.connected_region.name
-                    child_access_table[next_name] = True
-                    if next_name not in region_list:
-                        region_list.append(next_name)
-                    if next_name == Regions.TEMPLE_OF_TIME.value: # If we just got access to ToT, we need to reset the floodfill to cover the child side
-                        # If we're in this clause, we first gained access as adult, which must have been our starting age
-                        child_access_table[next_name] = True
-                        just_found_tot = True
-    return just_found_tot
+                if new_region in reachable:
+                    blocked.remove(connection)
+                elif connection.can_reach(self):
+                    reachable.add(new_region)
+                    blocked.remove(connection)
+                    blocked.update(new_region.exits)
+                    queue.extend(new_region.exits)
+                    self.path[new_region] = (new_region.name, self.path.get(connection, None))
+    
+    def _soh_can_reach_as_age(self, region: Regions, age: Ages, player: int): # Todo, type safety to age enum
+        if self._soh_age[player] is Ages.null:
+            # first layer of recursion
+            self._soh_age[player] = age
+            can_reach = self.multiworld.get_region(region.value, player).can_reach(self)
+            self._soh_age[player] = Ages.null
+            return can_reach
+        return self._soh_age[player] == age
 
-def can_access_region_as_child(state: CollectionState, world: "SohWorld", region: Region | str) -> bool:
-    if region is Region:
-        region = cast(Region, region).name
-    region = cast(str, region)
-    return region in child_access_table and child_access_table[region] and state.can_reach_region(region, world.player)
-
-def can_access_region_as_adult(state: CollectionState, world: "SohWorld", region: Region | str) -> bool:
-    if region is Region:
-        region = cast(Region, region).name
-    region = cast(str, region)
-    return region in adult_access_table and adult_access_table[region] and state.can_reach_region(region, world.player)
-
-def can_access_entrance_as_child( state: CollectionState, world: "SohWorld", entrance: Entrance | str) -> bool:
-    if isinstance(entrance, Entrance):
-        entrance = cast(Entrance, entrance)
-        if entrance.parent_region is None:
-            return False
-        return entrance.name in entrance_data_table and entrance_data_table[entrance.name].can_pass_as_child and state.can_reach_entrance(entrance.name, world.player)\
-              and can_access_region_as_child(state, world, entrance.parent_region)
-    else:
-        entrance = cast(str, entrance)
-        return can_access_entrance_as_child(state, world, world.get_entrance(entrance))
-
-def can_access_entrance_as_adult(state: CollectionState, world: "SohWorld", entrance: Entrance | str) -> bool:
-    if isinstance(entrance, Entrance):
-        entrance = cast(Entrance, entrance)
-        if entrance.parent_region is None:
-            return False
-        return entrance.name in entrance_data_table and entrance_data_table[entrance.name].can_pass_as_adult and state.can_reach_entrance(entrance.name, world.player) \
-              and can_access_region_as_adult(state, world, entrance.parent_region)
-    else:
-        entrance = cast(str, entrance)
-        return can_access_entrance_as_adult(state, world, world.get_entrance(entrance))
