+++ worlds/oot_soh/__init__.py
@@ -1,19 +1,17 @@
 from typing import Dict, Any
-from collections import deque
 
-from BaseClasses import CollectionState, Item, Tutorial, MultiWorld
+from BaseClasses import CollectionState, Item, Tutorial
 from Utils import visualize_regions
-from worlds.AutoWorld import WebWorld, World, LogicMixin
+from worlds.AutoWorld import WebWorld, World
 from .Items import SohItem, item_data_table, item_table, item_name_groups
 from .Locations import location_table
 from .Options import SohOptions
-from .RegionAgeAccess import reset_age_access, update_age_access
 from .Regions import create_regions_and_locations, place_locked_items
 from .Enums import *
 from .ItemPool import create_item_pool
+from . import RegionAgeAccess
 
 import logging
-import copy
 logger = logging.getLogger("SOH_OOT")
 
 class SohWebWorld(WebWorld):
@@ -30,80 +28,6 @@ class SohWebWorld(WebWorld):
     
     tutorials = [setup_en]
     game_info_languages = ["en"]
-
-class SohState(LogicMixin):
-    def init_mixin(self, parent: MultiWorld):
-        self._soh_stale = {player: True for player in parent.worlds.keys()
-                            if parent.worlds[player].game == SohWorld.game}
-        players = parent.get_game_groups(SohWorld.game) + parent.get_game_players(SohWorld.game)
-        self._soh_child_reachable_regions = {player: set() for player in players}
-        self._soh_adult_reachable_regions = {player: set() for player in players}
-        self._soh_child_blocked_regions = {player: set() for player in players}
-        self._soh_adult_blocked_regions = {player: set() for player in players}
-        self._soh_age = {player: Ages.null for player in players}
-
-    def copy_mixin(self, ret) -> CollectionState:
-        ret._soh_stale = {player: stale for player, stale in self._soh_stale.items()}
-        ret._soh_child_reachable_regions = {player: copy.copy(regions) for player, regions in self._soh_child_reachable_regions.items()}
-        ret._soh_adult_reachable_regions = {player: copy.copy(regions) for player, regions in self._soh_adult_reachable_regions.items()}
-        ret._soh_child_blocked_regions = {player: copy.copy(regions) for player, regions in self._soh_child_blocked_regions.items()}
-        ret._soh_adult_blocked_regions = {player: copy.copy(regions) for player, regions in self._soh_adult_blocked_regions.items()}
-        ret._soh_age = {player: age for player, age in self._soh_age.items()}
-        return ret
-    
-    def _soh_invalidate(self, player):
-        self._soh_child_reachable_regions[player] = set()
-        self._soh_adult_reachable_regions[player] = set()
-        self._soh_child_blocked_regions[player] = set()
-        self._soh_adult_blocked_regions[player] = set()
-        self._soh_stale[player] = True
-
-    def _soh_update_age_reachable_regions(self, player):
-        self._soh_stale[player] = False
-        for age in [Ages.CHILD, Ages.ADULT]:
-            self._soh_age[player] = age
-            start = self.multiworld.get_region(Regions.ROOT, player)
-            
-            if age == Ages.CHILD:
-                reachable = self._soh_child_reachable_regions[player]
-                blocked = self._soh_child_blocked_regions[player]
-                queue = deque(self._soh_child_blocked_regions[player])
-            else:
-                reachable = self._soh_adult_reachable_regions[player]
-                blocked = self._soh_adult_blocked_regions[player]
-                queue = deque(self._soh_adult_blocked_regions[player])
-
-            # init on first call
-            if start not in reachable:
-                reachable.add(start)
-                blocked.update(start.exits)
-                queue.extend(start.exits)
-
-            # run breadth first search
-            while queue:
-                connection = queue.popleft()
-                new_region = connection.connected_region
-                if new_region is None:
-                    continue
-                if new_region in reachable:
-                    blocked.remove(connection)
-                elif connection.can_reach(self):
-                    reachable.add(new_region)
-                    blocked.remove(connection)
-                    blocked.update(new_region.exits)
-                    queue.extend(new_region.exits)
-                    self.path[new_region] = (new_region.name, self.path.get(connection, None))
-    
-    def _soh_can_reach_as_age(self, region: Regions, age: Ages, player: int): # Todo, type safety to age enum
-        if self._soh_age[player] is Ages.null:
-            # first layer of recursion
-            self._soh_age[player] = age
-            can_reach = self.multiworld.get_region(region.value, player).can_reach(self)
-            self._soh_age[player] = Ages.null
-            return can_reach
-        return self._soh_age[player] == age
-
-
 
 class SohWorld(World):
     """A PC Port of Ocarina of Time"""
